// VIS_Application/app/Model/Masters/CurrencyRelated/currency.ts
export interface ICurrency {
    CompanyId: number;
    Id: number;
    Full_Name: string;
    Short_Name: string;
    Sub_Unit: string;
    Symbol: string;
    Is_Base_Currency: boolean;
    Default_Exchange: string;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;
    
}